Change Arch Linux site URL's
Recently Arch Linux shed the www portion of its site. The links still work,
but it's a 301 redirection to archlinux.org. Changed all the URL's still pointing
to www.archlinux.org to avoid the unnecessary redirection, since on browsers the 301
is usually saved and respected, but command line tools line wget or python might not
save this and always go through the redirect.