Update mkinitcpio.conf hooks for upstream consistency
This commit adjusts the hooks in mkinitcpio.conf to follow the changes
in the upstream mkinitcpio repository.
It also add the consolefont hook if udev is used instead of systemd.

Addresses upstream commits:
[mkinitcpio.conf: add kms to the default HOOKS array](https://gitlab.archlinux.org/archlinux/mkinitcpio/mkinitcpio/-/commit/b99eb1c0d5704ed0a0d220d0abcbbbeec10a3b76)
[mkinitcpio.conf: adjust the placement of the modconf hook for consistency](https://gitlab.archlinux.org/archlinux/mkinitcpio/mkinitcpio/-/commit/0b052b14615fefbb61ee3d813e3acd365332ee35)

Additionally, it replaces the missing `consolefont` hook if the
`keyboard` hook is used instead of the `sd-vconsole` hook.
For more details, see the Arch Wiki:
[dm-crypt/System configuration - 1.1 mkinitcpio](https://wiki.archlinux.org/title/Dm-crypt/System_configuration#mkinitcpio)