Refactor Argument Parsing Logic (#2605)
* Update Spanish translation for 'greeter'

* Moved the os.execve call below the debug logging statement in SysCommandWorker

* Fix call to removed function in DesktopProfile class

* Refactor: Move precondition check for '=' to the beginning of the loop

This commit moves the check for the '=' element to the start of the while loop in the `parse_unspecified_argument_list` function. This change improves code readability and ensures that the precondition is evaluated before processing other elements, enhancing the overall flow of the argument parsing logic.

* Refactor: Remove redundant 'else' statement in argument parsing logic

This commit simplifies the control flow in the `parse_unspecified_argument_list` function by removing a redundant 'else' block. The logic is streamlined to improve readability and maintainability while preserving the original functionality. The removal of unnecessary nesting helps clarify the conditions under which elements are processed.

* Refactor: Remove redundant 'else' statement in argument parsing logic

This commit simplifies the control flow in the `parse_unspecified_argument_list` function by removing a redundant 'else' block. The logic is streamlined to improve readability and maintainability while preserving the original functionality. The removal of unnecessary nesting helps clarify the conditions under which elements are processed.

* Replace conditional validation with filtering

This commit modifies the `parse_unspecified_argument_list` function to filter out unwanted elements from the `unknowns` list instead of using a conditional check with an `if` statement. This change improves code readability and efficiency by utilizing list comprehension to create a new list that excludes elements equal to "=".