docs: remove line block markup
I don't see a point in using [line blocks][1] anywhere in the
documentation. Note that mandoc consequently renders it with a `<pre>`
tag which looks ugly on [man.archlinux.org][2].

Also simplified the URL in `index.rst` which can be written verbatim
without any special markup.

See this [archmanweb issue][3] for details.

[1]: https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#line-blocks
[2]: https://man.archlinux.org/man/archinstall.1
[3]: https://gitlab.archlinux.org/archlinux/archmanweb/-/issues/49#note_213730