Updated translation and fixed errors in last translation. (#3515)
* still updating

* Updated urdu translation

Because it is very hard to display urdu fonts in consol. I changed urdu
writing to latin script which is also known as "Roman Urdu". For more
information check out https://en.wikipedia.org/wiki/Roman_Urdu.

* Updated errors and translation

As it is mentioned in #3463 the errors has been rectified and more
translation is added.